use primitive type in ClientUtils#getFormattedMinecraftTime

+++ src/main/java/net/geforcemods/securitycraft/util/ClientUtils.java
@@ -45,10 +45,10 @@ public class ClientUtils {
 	 * Returns the current Minecraft in-game time, in a 12-hour AM/PM format.
 	 */
 	public static String getFormattedMinecraftTime() {
-		Long time = Minecraft.getInstance().level.getDayTime();
-		int hours24 = (int) ((float) time.longValue() / 1000L + 6L) % 24;
+		long time = Minecraft.getInstance().level.getDayTime();
+		int hours24 = (int) ((float) time / 1000L + 6L) % 24;
 		int hours = hours24 % 12;
-		int minutes = (int) (time.longValue() / 16.666666F % 60.0F);
+		int minutes = (int) (time / 16.666666F % 60.0F);
 
 		return String.format("%02d:%02d %s", Integer.valueOf(hours < 1 ? 12 : hours), Integer.valueOf(minutes), hours24 < 12 ? "AM" : "PM");
 	}
